+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+Unreleased
+----------
+- Made `new_unchecked` constructor public
+
+
 0.1.6
 -----
 - Adjusted `Id::new` to panic when a `usize` counter overflow is